# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Sectra Reward Api",
        "version": "1.0.0"
    },
    "paths": {
        "/api/register": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Register Api",
                "description": "App Registration",
                "operationId": "authRegister",
                "requestBody": {
                    "description": "Pass user credentials",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "first_name",
                                    "last_name",
                                    "email",
                                    "mobile_no",
                                    "country_code",
                                    "gender",
                                    "dob"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "user1@mail.com"
                                    },
                                    "first_name": {
                                        "type": "string",
                                        "example": "John"
                                    },
                                    "last_name": {
                                        "type": "string",
                                        "example": "Son"
                                    },
                                    "mobile_no": {
                                        "type": "string",
                                        "example": "1234567890"
                                    },
                                    "country_code": {
                                        "type": "string",
                                        "example": "+1"
                                    },
                                    "gender": {
                                        "type": "string",
                                        "example": "Male"
                                    },
                                    "dob": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2022-01-01"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Validation Error"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/loginByMobileNumber": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Login By MobileNumber",
                "description": "Login by mobile number",
                "operationId": "authLogin",
                "requestBody": {
                    "description": "Pass user credentials",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "country_code",
                                    "mobile_no"
                                ],
                                "properties": {
                                    "country_code": {
                                        "type": "string",
                                        "example": "+91"
                                    },
                                    "mobile_no": {
                                        "type": "string",
                                        "example": "8435153945"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Validation Error"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/checkUser": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Check User MobileNumber",
                "description": "Check User mobile number",
                "operationId": "checkUser",
                "requestBody": {
                    "description": "Pass user credentials",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "country_code",
                                    "mobile_no"
                                ],
                                "properties": {
                                    "country_code": {
                                        "type": "string",
                                        "example": "+91"
                                    },
                                    "mobile_no": {
                                        "type": "string",
                                        "example": "8435153945"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Validation Error"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/category-list": {
            "post": {
                "tags": [
                    "Category"
                ],
                "summary": "Get list of category",
                "description": "Returns list of category",
                "operationId": "categoryList",
                "responses": {
                    "200": {
                        "description": "Category List",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "not found"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
        "/api/country-code": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Country code list",
                "description": "Country code list",
                "operationId": "countryCode",
                "responses": {
                    "200": {
                        "description": "Country code List",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "not found"
                    }
                }
            }
        },
        "/api/update-profile": {
            "post": {
                "tags": [
                    "Authorized Users"
                ],
                "summary": "Update Profile",
                "description": "Update Profile after register",
                "operationId": "updateProfile",
                "requestBody": {
                    "description": "Pass user credentials",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "emirates_id",
                                    "passport_no",
                                    "passport_expiry",
                                    "address_line1",
                                    "address_line2",
                                    "landmark",
                                    "pincode",
                                    "country",
                                    "state",
                                    "city"
                                ],
                                "properties": {
                                    "emirates_id": {
                                        "type": "string",
                                        "example": "TYYTU987979"
                                    },
                                    "passport_no": {
                                        "type": "string",
                                        "example": "678678786786"
                                    },
                                    "passport_expiry": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2024-12-12"
                                    },
                                    "address_line1": {
                                        "type": "string",
                                        "example": "Sayaji Hotel"
                                    },
                                    "address_line2": {
                                        "type": "string",
                                        "example": "Vijay Nagar"
                                    },
                                    "landmark": {
                                        "type": "string",
                                        "example": "Near C21 Mall"
                                    },
                                    "pincode": {
                                        "type": "string",
                                        "example": "452011"
                                    },
                                    "country": {
                                        "type": "string",
                                        "example": "India"
                                    },
                                    "state": {
                                        "type": "string",
                                        "example": "MP"
                                    },
                                    "city": {
                                        "type": "string",
                                        "example": "Indore"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Profile updated successfully",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "not found"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
        "/api/list-plans": {
            "post": {
                "tags": [
                    "Authorized Users"
                ],
                "summary": "Pricing Plans list",
                "description": "Pricing plans list",
                "operationId": "plansList",
                "responses": {
                    "200": {
                        "description": "Pricing plans",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "not found"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        }
    },
    "components": {
        "responses": {
            "200": {
                "description": "Register Successfully"
            }
        },
        "securitySchemes": {
            "bearer_token": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    }
}